Edited delete function to delete void pointers

+++ linked_lists.h
@@ -18,7 +18,7 @@ typedef struct LinkedList_t
     size_t dataSize;
 } LinkedList;
 
-LinkedList* ll_create(void* values[], const int n);
+LinkedList* ll_create(void* values[], const size_t n);
 int ll_insert(LinkedList* list, void* value, const int index);
 int ll_print(LinkedList* list, void print(void*));
 int ll_delete(LinkedList* list, const int index);
@@ -26,7 +26,7 @@ int ll_reverse(LinkedList* list);
 int ll_free(LinkedList* list);
 
 struct LibLinkedList_l {
-    LinkedList* (*create)(void* values[], const int n);
+    LinkedList* (*create)(void* values[], const size_t n);
     int (*insert)(LinkedList* list, void* value, const int index);
     int (*print)(LinkedList* list, void print(void*));
     int (*delete)(LinkedList* list, const int index);
